refactor: remove old about, nomepage, 404 pages
